move all data to /data/2015
