+++ data/2015/dataSources/CharacterStatDefinitions.json
