+++ data/2015/dataSources/EquipmentSlotDefinitions.json
