+++ src/servers/ZoneServer/commands/hax.ts
@@ -209,7 +209,7 @@ const hax: any = {
     server.sendChatText(client, "Delete player, back in observer mode");
   },
   changeStat: function (server: ZoneServer, client: Client, args: any[]) {
-    const stats = require("../../../../data/sampleData/stats.json");
+    const stats = require("../../../../data/2015/sampleData/stats.json");
     server.sendData(client, "PlayerUpdate.UpdateStat", {
       characterId: client.character.characterId,
       stats: stats,
@@ -306,7 +306,7 @@ const hax: any = {
           delete require.cache[
             require.resolve("../../../../data/sampleData/weather.json")
           ];
-          server._weatherTemplates = require("../../../../data/sampleData/weather.json");
+          server._weatherTemplates = require("../../../../data/2015/sampleData/weather.json");
         } else {
           await server._db?.collection("weathers").insertOne(currentWeather);
           server._weatherTemplates = await (server._db as any)
